requirements/production.txt
asgiref==3.2.7
boto3==1.13.1
botocore==1.16.1
django-admin-view-permission==1.9
django-filter==2.2.0
django-storages==1.9.1
django==3.0.5
docutils==0.15.2
jmespath==0.9.5
pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9; python_version != '3.4'
